Fixes condition for special-encoding fields
By simplifying access and modification of encoded fields we
introduced a regression in transform/check.

transform template function returns value in whatever format it received.
In this case base64.
When checking for result in Test we get base64("true") instead of just "true".

This change checks for base64 encoded strings as well.