Retry when Kinesis client is unable to start on a worker
We had to handle errors when a worker node is unable to spin up a working Kinesis
client. Previously we would report thes errors but the batch wouldn't be
properly handled.

This change adds retry mechanism that puts batches is retry directory and reruns
them after the main loading finishes. The reads are based upon Spark
Streaming[1] to automatically discover new files.

1 - https://spark.apache.org/docs/latest/structured-streaming-programming-guide.html