Add stable Flink kinesis job and deprecate Spark (close #132)
Spark Kinesis integration seems not to be very reliable. This reintroduces Flink
job that we are already using with Bad Row Aggregator.

Uses EMR 6.9.0 with Flink 1.15.2 that don't have Kinesis and EMR clustering
issues that were faced with EMR 6.10 and Flink 1.16.

Cloudwatch metrics are reported through StatsD protocol as shown with
reference configuration for Dataflow runner.