Handle url-unsafe Base64 encodings (close #137)
Some trackers base64-encode strings using url-unsafe methods. These
strings contain special 62nd and 63rd characters in alphabet: `-` and `_`.
This would previously lead to recovery failures as the string wasn't
properly decoded. And the error was attributed to the configuration issue.

In order to keep backwards-compatibility with the tracker behavior we need
to work around it by replacing the 62nd and 63rd characters with corresponding
characters from with URL and filename safe alphabet[1]: `+` and `/`.

1 - http://tools.ietf.org/html/rfc4648#page-7