Make NVP codec isomorphic
NVP (Name-Value-Pair) type casts all the source values into plain `String`.
This causes problems when a value is encoded differently and a different
type is required when recovering a failed event.

We need a smarter way of dealing with that. Ideally Name-Value-Pair should
be removed, but until then we need an alternative codec that will somehow
restore the type.