Do not decode iglu json webhook ue_pr paramter on coerce
When attempting convering RawEvent into CollectorPayload we previously
were convering it through proper JSON decoder. But as this is already
manually modified through recovery process we can skip decoding and
instead simply pass the JSON.
Otherwise we'd need to alter RawEvent to contain `body` field which it
is missing. And we cannot.